backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
langchain==0.1.0
openai==1.3.0
sentence-transformers==2.2.2
chromadb==0.4.18
PyMuPDF==1.23.8
python-docx==1.1.0
pypdf==3.17.1
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
apscheduler==3.10.4
python-dotenv==1.0.0
pydantic==2.5.0
selenium==4.15.2
webdriver-manager==4.0.1
paddlepaddle==2.5.2
paddleocr==2.7.0.3
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
huggingface-hub==0.19.4
cross-encoder==0.1.0
